moved the bot class into core
