+++ launcher.py
@@ -10,7 +10,7 @@ import matplotlib
 from loguru import logger
 
 import configurator as config
-from bot import Tomodachi
+from tomodachi.core.bot import Tomodachi
 from tomodachi.src.loguru_intercept import InterceptHandler
 from tomodachi.utils import cachemanager, psql
