+++ tomodachi/cogs/reminders.py
@@ -20,7 +20,7 @@ from discord import ConnectionClosed, Forbidden
 from discord.ext import commands
 from discord.ext.menus import MenuPages
 
-from bot import Tomodachi
+from tomodachi.core.bot import Tomodachi
 from tomodachi.src.myembed import MyEmbed
 from tomodachi.src.mymenus import MyPagesSource
 from tomodachi.utils import TimeInput
